rename: Move images folder public to assets
